+++ tests/imputers/test_MeanImputer.py
@@ -9,6 +9,7 @@ from tests.base_tests import (
     GenericFitTests,
     GenericTransformTests,
     OtherBaseBehaviourTests,
+    OtherBaseBehaviourTestsNumeric,
     ReturnNativeTests,
     WeightColumnFitMixinTests,
     WeightColumnInitMixinTests,
@@ -120,7 +121,9 @@ class TestLazyYSupport:
 
 
 class TestOtherBaseBehaviour(
-    OtherBaseBehaviourTests, EmptyColumnsFitTransformPassTests
+    OtherBaseBehaviourTests,
+    EmptyColumnsFitTransformPassTests,
+    OtherBaseBehaviourTestsNumeric,
 ):
     """
     Class to run tests for BaseTransformerBehaviour behaviour outside the three standard methods.