+++ tests/imputers/test_ModeImputer.py
@@ -10,6 +10,7 @@ from tests.base_tests import (
     GenericFitTests,
     GenericTransformTests,
     OtherBaseBehaviourTests,
+    OtherBaseBehaviourTestsNumeric,
     ReturnNativeTests,
     WeightColumnFitMixinTests,
     WeightColumnInitMixinTests,
@@ -313,7 +314,9 @@ class TestLazyYSupport:
 
 
 class TestOtherBaseBehaviour(
-    OtherBaseBehaviourTests, EmptyColumnsFitTransformPassTests
+    OtherBaseBehaviourTests,
+    EmptyColumnsFitTransformPassTests,
+    OtherBaseBehaviourTestsNumeric,
 ):
     """
     Class to run tests for BaseTransformerBehaviour outside the three standard methods.