+++ tests/mapping/test_MappingTransformer.py
@@ -8,6 +8,7 @@ from tests.mapping.test_BaseMappingTransformer import (
     BaseMappingTransformerTransformTests,
     GenericFitTests,
     OtherBaseBehaviourTests,
+    OtherBaseBehaviourTestsString,
 )
 from tests.utils import (
     _check_if_skip_test,
@@ -339,7 +340,9 @@ class TestTransform(BaseMappingTransformerTransformTests, ReturnNativeTests):
 
 
 class TestOtherBaseBehaviour(
-    OtherBaseBehaviourTests, EmptyMappingsFitTransformPassTests
+    OtherBaseBehaviourTests,
+    EmptyMappingsFitTransformPassTests,
+    OtherBaseBehaviourTestsString,
 ):
     """
     Class to run tests for BaseTransformerBehaviour outside the three standard methods.