+++ tests/nominal/test_GroupRareLevelsTransformer.py
@@ -15,6 +15,7 @@ from tests.base_tests import (
     EmptyColumnsFitTransformPassTests,
     GenericFitTests,
     OtherBaseBehaviourTests,
+    OtherBaseBehaviourTestsString,
     WeightColumnFitMixinTests,
     WeightColumnInitMixinTests,
 )
@@ -563,7 +564,9 @@ class TestLazyYSupport:
 
 
 class TestOtherBaseBehaviour(
-    OtherBaseBehaviourTests, EmptyColumnsFitTransformPassTests
+    OtherBaseBehaviourTests,
+    EmptyColumnsFitTransformPassTests,
+    OtherBaseBehaviourTestsString,
 ):
     """
     Class to run tests for BaseTransformerBehaviour outside the three standard methods.