+++ CHANGELOG.rst
@@ -27,13 +27,22 @@ We use the tags:
 
 Each individual change should have a link to the pull request after the description of the change.
 
-3.3.0 (unreleased)
+3.4.0 (30/04/2026)
 ------------------
 
 Changed
 ^^^^^^^
 - added LowerCaseTransformer and RemoveCharactersTransformer to string module
 - added functions module for stateless transforms
+
+3.3.0 (30/04/2026)
+------------------
+
+Changed
+^^^^^^^
+- feat: added `check_is_fitted` to all transformers `#494 <https://github.com/azukds/tubular/issues/494>`_
+- chore: added python 3.14 on automated testing `#494 <https://github.com/azukds/tubular/issues/494>`_
+- converted MeanResponseTransformer to work with lazyframes `#533 <https://github.com/azukds/tubular/issues/533>_`
 - updated pytest version on dependabot recommendation
 
 3.2.0 (14/04/2026)